gwpy.io.nds2: renamed module
nds2 is the protocol, so its a more precise name
