+++ gwpy/io/nds2.py
@@ -20,7 +20,7 @@
 to LIGO data.
 """
 
-from __future__ import print_function
+from __future__ import (absolute_import, print_function)
 
 import os
 import sys
